requirements.txt
asgiref==3.8.1
Django==5.2
django-dotenv==1.4.2
djangorestframework==3.16.0
dotenv==0.9.9
celery==5.5.1
requests==2.32.3
drf-spectacular==0.28.0






